Create Chat_Redirector.xml
Added the chat redirector plugin [created by Nick Gammon, modified by me to interpret the MUD-specific channels]. Note that between my unofficial release of this modified script and now, I've removed "say" from the capture clause, as many of the NPCs in the mud have "say" commands which are better placed in the main "action" window to prevent confusion. "Says" are here and now and typically not where MUD users have conversations.